Marqueeをjsonで要素を変更できるようにしました．
