+++ features/main/components/index.tsx
@@ -1,8 +1,7 @@
 import GitCalendar from "./GitCalendar/GitCalendar";
 import Hobby from "./Hobby/Hobby";
-import MarqueeWidget from "./Marquee/Marquee";
 import Notice from "./Notice/Notice";
 import Profile from "./Profile/Profile";
 import Welcome from "./Welcome/Welcome";
 
-export { GitCalendar, Hobby, MarqueeWidget, Notice, Profile, Welcome };
+export { GitCalendar, Hobby, Notice, Profile, Welcome };
